[FIX] Helpdesk Management Timesheet
Before this commit, in any view with ticket_id present but unused the onchange would be called and set project and task to False. It was found while trying to integrate project_timesheet_time_control with helpdesk_mgmt_timesheet, where this onchange overrides the results of default_get.

After this commit, if a helpdesk ticket is not present when calling onchange, project and task ids are not reset to False.