[IMP] Make changing ticket type less annoying
Before this commit changing ticket type on a helpdesk ticket would
indiscriminately set the team and user to NULL. This was very annoying
especially for tickets received via email as it would wipe preset
fields.

After this commit, team and user is only set to NULL if it isn't a
valid team for the ticket type. If there is only 1 valid team the
helpdesk team is changed to that.