Fixed permission error
When trying to open a res_partner form with a user who doesn't have permission for read fieldservice data, this view causes an error cause it try to set attributes to a button that doesn't exist. Specifying groups_id fix this, extending the view only if the current user has permissions to see that button.