[IMP] Restore the fields in the ticket creation
The way it worked until the change was that these fields were populated with the customer's information when the ticket was created.
And to preserve that functionality I think it is necessary to keep informing those fields when creating the ticket from the portal.
Because when creating the ticket from the back-end the behavior is similar, when selecting a customer, the fields are filled automatically.