[IMP] helpdesk_mgmt: Delete unused record
In the commit https://github.com/OCA/helpdesk/commit/4382b1ba944e41ca6e2c95340fbe53c8269a4bd3
the record assignment_email_template was deleted but in the migration it was restored but it
is not used anymore.

So we want to avoid this record creation and remove it from database.