[IMP] helpdesk_mgmt_sale: Access tickets from the order
With the widget ‘many2many_tags’ it is not possible to click on a ticket
in the order and access the form. By adding a smartbutton it is possible
to access the list of tickets in the order and also consult the form for
each one of them.

TT50683

[UPD] Update helpdesk_mgmt_sale.pot

[BOT] post-merge updates

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: helpdesk-15.0/helpdesk-15.0-helpdesk_mgmt_sale
Translate-URL: https://translation.odoo-community.org/projects/helpdesk-15-0/helpdesk-15-0-helpdesk_mgmt_sale/