[FIX] helpdesk_mgmt: company_id must always be available for check_company
In a previous commit check_company=True was added to the partner_id field.
company_id was added as a hidden field to the form view. Unfortunately
company_id was already present for users having multi-company rights. When
updating modules based on helpdesk_mgmt this leads to errors. This should not
happen because of the hidden field, but adding it this way prevents the
error.