[FIX] helpdesk_mgmt: Don't reset stage on user change
Steps to reproduce:

- Create a ticket in a specific team.
- Change the stage to one that is not the first.
- Assign user of the same team.

Result: the stage is reset to the first one.

Analysis: Even if the team is not changed, the `_compute_stage_id`
method is executed. We don't want to remove computed writable methods,
as they are very useful, so let's just do a preventive check on the
compute method for not changing the stage if the current one is still
applicable for the current team.

It includes a test for avoiding future regressions.